redir parsing corrected dealing with quotes
