--- heredoc_output.txt
@@ -1,6 +1,0 @@
-
-
-
-
-
-
