+++ main.c
@@ -6,20 +6,11 @@
 /*   By: gecarval <gecarval@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 08:31:04 by gecarval          #+#    #+#             */
-/*   Updated: 2024/11/21 16:33:26 by gecarval         ###   ########.fr       */
+/*   Updated: 2024/11/27 16:39:07 by gecarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "./includes/minishell.h"
-
-void	ft_signal_handler(int signum)
-{
-	(void)signum;
-	write(1, "\n", 1);
-	rl_on_new_line();
-	rl_replace_line("", 0);
-	rl_redisplay();
-}
 
 // This function prints the command structure List for debugging purposes
 void	ft_print_cmd(t_cmd *cmd)
@@ -42,19 +33,13 @@ void	ft_print_cmd(t_cmd *cmd)
 		if (tmp->fd.filename_in != NULL)
 			printf("filename_in: %s\n", tmp->fd.filename_in);
 		else
-			printf("filename_in: %s\n",  "(null)");
+			printf("filename_in: %s\n", "(null)");
 		if (tmp->fd.filename_out != NULL)
 			printf("filename_out: %s\n", tmp->fd.filename_out);
 		else
 			printf("filename_out: %s\n", "(null)");
 		tmp = tmp->next;
 	}
-}
-
-void	ft_default_fd(t_shell *shell)
-{
-	shell->fd_in = STDIN_FILENO;
-	shell->fd_out = STDOUT_FILENO;
 }
 
 // This start by initializing shell data struct
@@ -81,14 +66,17 @@ int	main(int argc, char **argv, char **envp)
 			write(1, "exit\n", 5);
 		if (shell.line == NULL)
 			break ;
-		add_history(shell.line);
-		parse_line(&shell);
-		ft_print_cmd(shell.cmd);
-		exec_cmd(&shell);
-		free_cmd(&shell.cmd);
-		if (shell.line != NULL)
-			free(shell.line);
-		ft_default_fd(&shell);
+		if (shell.line[0] != '\0')
+		{
+			add_history(shell.line);
+			parse_line(&shell);
+			ft_print_cmd(shell.cmd);
+			exec_cmd(&shell);
+			free_cmd(&shell.cmd);
+			if (shell.line != NULL)
+				free(shell.line);
+			ft_default_fd(&shell);
+		}
 	}
 	ft_free_all(&shell);
 	return (0);
